+++ aeneas/extra/ctw_speect.py
@@ -36,11 +36,11 @@ import speect.audio_riff
 from aeneas.audiofile import AudioFile
 from aeneas.language import Language
 from aeneas.timevalue import TimeValue
-from aeneas.ttswrapper import TTSWrapper
+from aeneas.ttswrappers.basettswrapper import BaseTTSWrapper
 import aeneas.globalfunctions as gf
 
 
-class CustomTTSWrapper(TTSWrapper):
+class CustomTTSWrapper(BaseTTSWrapper):
     """
     A wrapper for the ``speect`` TTS engine.
 
@@ -75,9 +75,9 @@ class CustomTTSWrapper(TTSWrapper):
 
     #
     # NOTE in this example we load a voice producing
-    #      audio data in PCM16 mono WAVE (RIFF) format
+    #      audio data in PCM16 mono WAVE (RIFF) format at 16000 Hz
     #
-    OUTPUT_MONO_WAVE = True
+    OUTPUT_AUDIO_FORMAT = ("pcm_s16le", 1, 16000)
 
     def __init__(self, rconf=None, logger=None):
         super(CustomTTSWrapper, self).__init__(