+++ aeneas/tests/test_espeakttswrapper.py
@@ -23,14 +23,14 @@
 
 import unittest
 
-from aeneas.espeakwrapper import ESPEAKWrapper
 from aeneas.textfile import TextFile
 from aeneas.textfile import TextFragment
+from aeneas.ttswrappers.espeakttswrapper import ESPEAKTTSWrapper
 from aeneas.runtimeconfiguration import RuntimeConfiguration
 import aeneas.globalfunctions as gf
 
 
-class TestESPEAKWrapper(unittest.TestCase):
+class TestESPEAKTTSWrapper(unittest.TestCase):
 
     def synthesize_single(self, text, language, ofp=None, zero_length=False):
         def inner(c_ext, cew_subprocess):
@@ -43,7 +43,7 @@ class TestESPEAKWrapper(unittest.TestCase):
                 rconf = RuntimeConfiguration()
                 rconf[RuntimeConfiguration.C_EXTENSIONS] = c_ext
                 rconf[RuntimeConfiguration.CEW_SUBPROCESS_ENABLED] = cew_subprocess
-                tts_engine = ESPEAKWrapper(rconf=rconf)
+                tts_engine = ESPEAKTTSWrapper(rconf=rconf)
                 result = tts_engine.synthesize_single(text, language, output_file_path)
                 gf.delete_file(handler, output_file_path)
                 if zero_length:
@@ -68,7 +68,7 @@ class TestESPEAKWrapper(unittest.TestCase):
                 rconf = RuntimeConfiguration()
                 rconf[RuntimeConfiguration.C_EXTENSIONS] = c_ext
                 rconf[RuntimeConfiguration.CEW_SUBPROCESS_ENABLED] = cew_subprocess
-                tts_engine = ESPEAKWrapper(rconf=rconf)
+                tts_engine = ESPEAKTTSWrapper(rconf=rconf)
                 anchors, total_time, num_chars = tts_engine.synthesize_multiple(
                     text_file,
                     output_file_path,
@@ -98,69 +98,69 @@ class TestESPEAKWrapper(unittest.TestCase):
             self.synthesize_multiple(None, zero_length=True)
 
     def test_multiple_invalid_output_path(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG, [u"word"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG, [u"word"])])
         with self.assertRaises(OSError):
             self.synthesize_multiple(tfl, ofp="x/y/z/not_existing.wav")
 
     def test_multiple_no_fragments(self):
         tfl = TextFile()
-        tfl.set_language(ESPEAKWrapper.ENG)
+        tfl.set_language(ESPEAKTTSWrapper.ENG)
         with self.assertRaises(ValueError):
             self.synthesize_multiple(tfl)
 
     def test_multiple_unicode_ascii(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG, [u"word"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG, [u"word"])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_unicode_unicode(self):
-        tfl = self.tfl([(ESPEAKWrapper.DEU, [u"Ausführliche"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.DEU, [u"Ausführliche"])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_empty(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG, [u""])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG, [u""])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_empty_multiline(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG, [u"", u"", u""])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG, [u"", u"", u""])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_empty_fragments(self):
         tfl = self.tfl([
-            (ESPEAKWrapper.ENG, [u""]),
-            (ESPEAKWrapper.ENG, [u""]),
-            (ESPEAKWrapper.ENG, [u""]),
+            (ESPEAKTTSWrapper.ENG, [u""]),
+            (ESPEAKTTSWrapper.ENG, [u""]),
+            (ESPEAKTTSWrapper.ENG, [u""]),
         ])
         self.synthesize_multiple(tfl)
 
     def test_multiple_empty_mixed(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG, [u"Word", u"", u"Word"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG, [u"Word", u"", u"Word"])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_empty_mixed_fragments(self):
         tfl = self.tfl([
-            (ESPEAKWrapper.ENG, [u"Word"]),
-            (ESPEAKWrapper.ENG, [u""]),
-            (ESPEAKWrapper.ENG, [u"Word"]),
+            (ESPEAKTTSWrapper.ENG, [u"Word"]),
+            (ESPEAKTTSWrapper.ENG, [u""]),
+            (ESPEAKTTSWrapper.ENG, [u"Word"]),
         ])
         self.synthesize_multiple(tfl)
 
     def test_multiple_replace_language(self):
-        tfl = self.tfl([(ESPEAKWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"])])
         self.synthesize_multiple(tfl)
 
     def test_multiple_replace_language_mixed(self):
         tfl = self.tfl([
-            (ESPEAKWrapper.UKR, [u"Word"]),
-            (ESPEAKWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"]),
-            (ESPEAKWrapper.UKR, [u"Word"])
+            (ESPEAKTTSWrapper.UKR, [u"Word"]),
+            (ESPEAKTTSWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"]),
+            (ESPEAKTTSWrapper.UKR, [u"Word"])
         ])
         self.synthesize_multiple(tfl)
 
     def test_multiple_replace_language_mixed_fragments(self):
         tfl = self.tfl([
-            (ESPEAKWrapper.ENG, [u"Word"]),
-            (ESPEAKWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"]),
-            (ESPEAKWrapper.ENG, [u"Word"])
+            (ESPEAKTTSWrapper.ENG, [u"Word"]),
+            (ESPEAKTTSWrapper.UKR, [u"Временами Сашке хотелось перестать делать то"]),
+            (ESPEAKTTSWrapper.ENG, [u"Word"])
         ])
         self.synthesize_multiple(tfl)
 
@@ -170,39 +170,39 @@ class TestESPEAKWrapper(unittest.TestCase):
             self.synthesize_multiple(tfl)
 
     def test_multiple_variation_language(self):
-        tfl = self.tfl([(ESPEAKWrapper.ENG_GBR, [u"Word"])])
+        tfl = self.tfl([(ESPEAKTTSWrapper.ENG_GBR, [u"Word"])])
         self.synthesize_multiple(tfl)
 
     def test_single_none(self):
         with self.assertRaises(TypeError):
-            self.synthesize_single(None, ESPEAKWrapper.ENG)
+            self.synthesize_single(None, ESPEAKTTSWrapper.ENG)
 
     def test_single_invalid_output_path(self):
         with self.assertRaises(OSError):
-            self.synthesize_single(u"word", ESPEAKWrapper.ENG, ofp="x/y/z/not_existing.wav")
+            self.synthesize_single(u"word", ESPEAKTTSWrapper.ENG, ofp="x/y/z/not_existing.wav")
 
     def test_single_empty_string(self):
-        self.synthesize_single(u"", ESPEAKWrapper.ENG, zero_length=True)
+        self.synthesize_single(u"", ESPEAKTTSWrapper.ENG, zero_length=True)
 
     def test_single_text_str_ascii(self):
         with self.assertRaises(TypeError):
-            self.synthesize_single(b"Word", ESPEAKWrapper.ENG)
+            self.synthesize_single(b"Word", ESPEAKTTSWrapper.ENG)
 
     def test_single_text_str_unicode(self):
         with self.assertRaises(TypeError):
-            self.synthesize_single(b"Ausf\xc3\xbchrliche", ESPEAKWrapper.DEU)
+            self.synthesize_single(b"Ausf\xc3\xbchrliche", ESPEAKTTSWrapper.DEU)
 
     def test_single_text_unicode_ascii(self):
-        self.synthesize_single(u"Word", ESPEAKWrapper.ENG)
+        self.synthesize_single(u"Word", ESPEAKTTSWrapper.ENG)
 
     def test_single_text_unicode_unicode(self):
-        self.synthesize_single(u"Ausführliche", ESPEAKWrapper.DEU)
+        self.synthesize_single(u"Ausführliche", ESPEAKTTSWrapper.DEU)
 
     def test_single_variation_language(self):
-        self.synthesize_single(u"Word", ESPEAKWrapper.ENG_GBR)
+        self.synthesize_single(u"Word", ESPEAKTTSWrapper.ENG_GBR)
 
     def test_single_replace_language(self):
-        self.synthesize_single(u"Временами Сашке хотелось перестать делать то", ESPEAKWrapper.UKR)
+        self.synthesize_single(u"Временами Сашке хотелось перестать делать то", ESPEAKTTSWrapper.UKR)
 
     def test_single_invalid_language(self):
         with self.assertRaises(ValueError):